Extract errors into their own files
