+++ lib/cursor_pager.rb
@@ -1,10 +1,15 @@
 # frozen_string_literal: true
 
 require "cursor_pager/version"
-require "cursor_pager/page"
-require "cursor_pager/order_value"
 
 module CursorPager
-  class Error < StandardError; end
-  # Your code goes here...
+  class Error < StandardError
+  end
 end
+
+require "cursor_pager/conflicting_orders_error"
+require "cursor_pager/cursor_not_found_error"
+require "cursor_pager/order_value_error"
+
+require "cursor_pager/order_value"
+require "cursor_pager/page"
