+++ lib/cursor_pager/cursor_not_found_error.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module CursorPager
+  class CursorNotFoundError < Error
+    def initialize(cursor)
+      message = "Couldn't find item for cursor #{cursor}."
+
+      super(message)
+    end
+  end
+end
