+++ lib/cursor_pager/order_value_error.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module CursorPager
+  class OrderValueError < Error
+  end
+end
