+++ lib/cursor_pager/page.rb
@@ -1,25 +1,6 @@
 # frozen_string_literal: true
 
 module CursorPager
-  class CursorNotFoundError < StandardError
-    def initialize(cursor)
-      message = "Couldn't find item for cursor #{cursor}."
-
-      super(message)
-    end
-  end
-
-  class ConflictingOrdersError < StandardError
-    MESSAGE = <<~MESSAGE
-      Ordering by multiple attributes requires they are all ordered in the
-      same direction.
-    MESSAGE
-
-    def initialize
-      super(MESSAGE)
-    end
-  end
-
   class Page
     attr_reader :relation, :first, :last, :after, :before, :order_values
